api-server/app/core/security.py
from datetime import datetime, timedelta
from typing import Optional
import hashlib
import base64
import bcrypt
from jose import JWTError, jwt

from app.core.config import settings

SECRET_KEY = settings.SECRET_KEY
ALGORITHM = settings.ALGORITHM
ACCESS_TOKEN_EXPIRE_MINUTES = settings.ACCESS_TOKEN_EXPIRE_MINUTES

def _prehash(password: str) -> bytes:
    """Pre-hash password with SHA-256 to safely handle bcrypt's 72-byte limit."""
    hashed = hashlib.sha256(password.encode("utf-8")).digest()
    return base64.b64encode(hashed)

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return bcrypt.checkpw(
        _prehash(plain_password),
        hashed_password.encode("utf-8")
    )

def get_password_hash(password: str) -> str:
    return bcrypt.hashpw(
        _prehash(password),
        bcrypt.gensalt()
    ).decode("utf-8")

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
