api-server/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
pydantic-settings

pydantic
httpx
bcrypt
python-jose[cryptography]
python-multipart
email-validator
